move test from dataview to common test files
