+++ test/unit/middleware/common.middleware.test.js
@@ -1,340 +1,382 @@
 import { describe, it, expect, vi } from 'vitest'
-import { createPaginationTemplateParams, addDatabaseFieldToSpecification, replaceUnderscoreInSpecification, pullOutDatasetSpecification, extractJsonFieldFromEntities, replaceUnderscoreInEntities, setDefaultParams, getIsPageNumberInRange } from '../../../src/middleware/common.middleware'
+import { createPaginationTemplateParams, addDatabaseFieldToSpecification, replaceUnderscoreInSpecification, pullOutDatasetSpecification, extractJsonFieldFromEntities, replaceUnderscoreInEntities, setDefaultParams, getIsPageNumberInRange, getUniqueDatasetFieldsFromSpecification } from '../../../src/middleware/common.middleware'
 import logger from '../../../src/utils/logger'
 
-describe('common.middleware.test.js', () => {
-  describe('createPaginationTemplateParams', () => {
-    it('creates pagination object with correct parameters', () => {
-      const req = {
-        resultsCount: 100,
-        urlSubPath: '/api/results/',
-        paginationPageLength: 10,
-        params: { pageNumber: 2 }
-      }
-      const res = {}
-      const next = vi.fn()
+describe('createPaginationTemplateParams', () => {
+  it('creates pagination object with correct parameters', () => {
+    const req = {
+      resultsCount: 100,
+      urlSubPath: '/api/results/',
+      paginationPageLength: 10,
+      params: { pageNumber: 2 }
+    }
+    const res = {}
+    const next = vi.fn()
 
-      createPaginationTemplateParams(req, res, next)
+    createPaginationTemplateParams(req, res, next)
 
-      expect(req.pagination).toEqual({
-        previous: { href: '/api/results/1' },
-        next: { href: '/api/results/3' },
-        items: [
-          { type: 'number', number: 1, href: '/api/results/1', current: false },
-          { type: 'number', number: 2, href: '/api/results/2', current: true },
-          { type: 'number', number: 3, href: '/api/results/3', current: false },
-          { type: 'ellipsis', ellipsis: true, href: '#' },
-          { type: 'number', number: 10, href: '/api/results/10', current: false }
+    expect(req.pagination).toEqual({
+      previous: { href: '/api/results/1' },
+      next: { href: '/api/results/3' },
+      items: [
+        { type: 'number', number: 1, href: '/api/results/1', current: false },
+        { type: 'number', number: 2, href: '/api/results/2', current: true },
+        { type: 'number', number: 3, href: '/api/results/3', current: false },
+        { type: 'ellipsis', ellipsis: true, href: '#' },
+        { type: 'number', number: 10, href: '/api/results/10', current: false }
+      ]
+    })
+    expect(next).toHaveBeenCalledTimes(1)
+  })
+
+  it('handles edge cases for pagination', () => {
+    const req = {
+      resultsCount: 10,
+      urlSubPath: '/api/results/',
+      paginationPageLength: 10,
+      params: { pageNumber: 1 }
+    }
+    const res = {}
+    const next = vi.fn()
+
+    createPaginationTemplateParams(req, res, next)
+
+    expect(req.pagination).toEqual({
+      items: [
+        { type: 'number', number: 1, href: '/api/results/1', current: true }
+      ]
+    })
+    expect(next).toHaveBeenCalledTimes(1)
+  })
+
+  it('handles invalid page numbers (negative)', () => {
+    const req = {
+      resultsCount: 100,
+      urlSubPath: '/api/results/',
+      paginationPageLength: 10,
+      params: { pageNumber: -1 }
+    }
+    const res = {}
+    const next = vi.fn()
+
+    expect(() => createPaginationTemplateParams(req, res, next)).toThrowError('Invalid page number')
+
+    expect(next).not.toHaveBeenCalled()
+  })
+
+  it('handles invalid page numbers (non-numeric)', () => {
+    const req = {
+      resultsCount: 100,
+      urlSubPath: '/api/results/',
+      paginationPageLength: 10,
+      params: { pageNumber: 'abc' }
+    }
+    const res = {}
+    const next = vi.fn()
+
+    expect(() => createPaginationTemplateParams(req, res, next)).toThrowError('Invalid page number')
+
+    expect(next).not.toHaveBeenCalled()
+  })
+
+  it('handles zero total results', () => {
+    const req = {
+      resultsCount: 0,
+      urlSubPath: '/api/results/',
+      paginationPageLength: 10,
+      params: { pageNumber: 1 }
+    }
+    const res = {}
+    const next = vi.fn()
+
+    createPaginationTemplateParams(req, res, next)
+
+    expect(req.pagination).toBeUndefined()
+    expect(next).toHaveBeenCalledTimes(1)
+  })
+
+  it('handles page number beyond available pages', () => {
+    const req = {
+      resultsCount: 50,
+      urlSubPath: '/api/results/',
+      paginationPageLength: 10,
+      params: { pageNumber: 6 }
+    }
+    const res = {}
+    const next = vi.fn()
+
+    createPaginationTemplateParams(req, res, next)
+
+    expect(req.pagination).toEqual({
+      previous: { href: '/api/results/5' },
+      items: [
+        { type: 'number', number: 1, href: '/api/results/1', current: false },
+        { type: 'number', number: 2, href: '/api/results/2', current: false },
+        { type: 'number', number: 3, href: '/api/results/3', current: false },
+        { type: 'number', number: 4, href: '/api/results/4', current: false },
+        { type: 'number', number: 5, href: '/api/results/5', current: false }
+      ]
+    })
+    expect(next).toHaveBeenCalledTimes(1)
+  })
+})
+
+describe('replaceUnderscoreInSpecification', () => {
+  it('replaces underscore with hyphen in specification datasetFields', () => {
+    const req = {
+      specification: {
+        fields: [
+          { datasetField: 'dataset_field' },
+          { datasetField: 'data_set_field' }
         ]
-      })
-      expect(next).toHaveBeenCalledTimes(1)
-    })
-
-    it('handles edge cases for pagination', () => {
-      const req = {
-        resultsCount: 10,
-        urlSubPath: '/api/results/',
-        paginationPageLength: 10,
-        params: { pageNumber: 1 }
       }
-      const res = {}
-      const next = vi.fn()
+    }
+    const res = {}
+    const next = vi.fn()
 
-      createPaginationTemplateParams(req, res, next)
+    replaceUnderscoreInSpecification(req, res, next)
 
-      expect(req.pagination).toEqual({
-        items: [
-          { type: 'number', number: 1, href: '/api/results/1', current: true }
+    expect(req.specification.fields).toHaveLength(2)
+    expect(req.specification.fields).toEqual([
+      { datasetField: 'dataset-field' },
+      { datasetField: 'data-set-field' }
+    ])
+  })
+
+  it('calls next function', () => {
+    const req = {
+      specification: {
+        fields: [
+          { datasetField: 'dataset_field' },
+          { datasetField: 'data_set_field' }
         ]
-      })
-      expect(next).toHaveBeenCalledTimes(1)
-    })
-
-    it('handles invalid page numbers (negative)', () => {
-      const req = {
-        resultsCount: 100,
-        urlSubPath: '/api/results/',
-        paginationPageLength: 10,
-        params: { pageNumber: -1 }
       }
-      const res = {}
-      const next = vi.fn()
+    }
+    const res = {}
+    const next = vi.fn()
 
-      expect(() => createPaginationTemplateParams(req, res, next)).toThrowError('Invalid page number')
+    replaceUnderscoreInSpecification(req, res, next)
+    expect(next).toHaveBeenCalledTimes(1)
+  })
 
-      expect(next).not.toHaveBeenCalled()
-    })
-
-    it('handles invalid page numbers (non-numeric)', () => {
-      const req = {
-        resultsCount: 100,
-        urlSubPath: '/api/results/',
-        paginationPageLength: 10,
-        params: { pageNumber: 'abc' }
-      }
-      const res = {}
-      const next = vi.fn()
-
-      expect(() => createPaginationTemplateParams(req, res, next)).toThrowError('Invalid page number')
-
-      expect(next).not.toHaveBeenCalled()
-    })
-
-    it('handles zero total results', () => {
-      const req = {
-        resultsCount: 0,
-        urlSubPath: '/api/results/',
-        paginationPageLength: 10,
-        params: { pageNumber: 1 }
-      }
-      const res = {}
-      const next = vi.fn()
-
-      createPaginationTemplateParams(req, res, next)
-
-      expect(req.pagination).toBeUndefined()
-      expect(next).toHaveBeenCalledTimes(1)
-    })
-
-    it('handles page number beyond available pages', () => {
-      const req = {
-        resultsCount: 50,
-        urlSubPath: '/api/results/',
-        paginationPageLength: 10,
-        params: { pageNumber: 6 }
-      }
-      const res = {}
-      const next = vi.fn()
-
-      createPaginationTemplateParams(req, res, next)
-
-      expect(req.pagination).toEqual({
-        previous: { href: '/api/results/5' },
-        items: [
-          { type: 'number', number: 1, href: '/api/results/1', current: false },
-          { type: 'number', number: 2, href: '/api/results/2', current: false },
-          { type: 'number', number: 3, href: '/api/results/3', current: false },
-          { type: 'number', number: 4, href: '/api/results/4', current: false },
-          { type: 'number', number: 5, href: '/api/results/5', current: false }
+  it('handles entities with no underscore in keys', () => {
+    const req = {
+      specification: {
+        fields: [
+          { datasetField: 'datasetField' }
         ]
-      })
-      expect(next).toHaveBeenCalledTimes(1)
+      }
+    }
+    const res = {}
+    const next = vi.fn()
+
+    replaceUnderscoreInSpecification(req, res, next)
+    expect(req.specification.fields).toHaveLength(1)
+    expect(req.specification.fields[0]).toEqual({ datasetField: 'datasetField' })
+  })
+})
+
+describe('addDatabaseFieldToSpecification', () => {
+  it('adds database field to specification fields', () => {
+    const req = {
+      specification: {
+        fields: [
+          { field: 'name' },
+          { field: 'address' }
+        ]
+      },
+      fieldMappings: [
+        { field: 'name', replacement_field: 'full_name' },
+        { field: 'address', replacement_field: 'physical_address' }
+      ]
+    }
+    const res = {}
+    const next = vi.fn()
+
+    addDatabaseFieldToSpecification(req, res, next)
+    expect(req.specification.fields).toHaveLength(2)
+    expect(req.specification.fields[0]).toEqual({
+      field: 'name',
+      datasetField: 'full_name'
+    })
+    expect(req.specification.fields[1]).toEqual({
+      field: 'address',
+      datasetField: 'physical_address'
     })
   })
 
-  describe('replaceUnderscoreInSpecification', () => {
-    it('replaces underscore with hyphen in specification datasetFields', () => {
-      const req = {
-        specification: {
-          fields: [
-            { datasetField: 'dataset_field' },
-            { datasetField: 'data_set_field' }
-          ]
-        }
-      }
-      const res = {}
-      const next = vi.fn()
+  it('handles special case for GeoX and GeoY fields', () => {
+    const req = {
+      specification: {
+        fields: [
+          { field: 'GeoX' },
+          { field: 'GeoY' }
+        ]
+      },
+      fieldMappings: []
+    }
+    const res = {}
+    const next = vi.fn()
 
-      replaceUnderscoreInSpecification(req, res, next)
-
-      expect(req.specification.fields).toHaveLength(2)
-      expect(req.specification.fields).toEqual([
-        { datasetField: 'dataset-field' },
-        { datasetField: 'data-set-field' }
-      ])
+    addDatabaseFieldToSpecification(req, res, next)
+    expect(req.specification.fields).toHaveLength(2)
+    expect(req.specification.fields[0]).toEqual({
+      field: 'GeoX',
+      datasetField: 'point'
     })
-
-    it('calls next function', () => {
-      const req = {
-        specification: {
-          fields: [
-            { datasetField: 'dataset_field' },
-            { datasetField: 'data_set_field' }
-          ]
-        }
-      }
-      const res = {}
-      const next = vi.fn()
-
-      replaceUnderscoreInSpecification(req, res, next)
-      expect(next).toHaveBeenCalledTimes(1)
-    })
-
-    it('handles entities with no underscore in keys', () => {
-      const req = {
-        specification: {
-          fields: [
-            { datasetField: 'datasetField' }
-          ]
-        }
-      }
-      const res = {}
-      const next = vi.fn()
-
-      replaceUnderscoreInSpecification(req, res, next)
-      expect(req.specification.fields).toHaveLength(1)
-      expect(req.specification.fields[0]).toEqual({ datasetField: 'datasetField' })
+    expect(req.specification.fields[1]).toEqual({
+      field: 'GeoY',
+      datasetField: 'point'
     })
   })
 
-  describe('addDatabaseFieldToSpecification', () => {
-    it('adds database field to specification fields', () => {
-      const req = {
-        specification: {
-          fields: [
-            { field: 'name' },
-            { field: 'address' }
-          ]
-        },
-        fieldMappings: [
-          { field: 'name', replacement_field: 'full_name' },
-          { field: 'address', replacement_field: 'physical_address' }
+  it('handles single GeoX field without GeoY', () => {
+    const req = {
+      specification: {
+        fields: [
+          { field: 'GeoX' }
+        ]
+      },
+      fieldMappings: []
+    }
+    const res = {}
+    const next = vi.fn()
+
+    addDatabaseFieldToSpecification(req, res, next)
+    expect(req.specification.fields).toHaveLength(1)
+    expect(req.specification.fields[0]).toEqual({
+      field: 'GeoX',
+      datasetField: 'point'
+    })
+  })
+
+  it('handles single GeoY field without GeoX', () => {
+    const req = {
+      specification: {
+        fields: [
+          { field: 'GeoY' }
+        ]
+      },
+      fieldMappings: []
+    }
+    const res = {}
+    const next = vi.fn()
+
+    addDatabaseFieldToSpecification(req, res, next)
+    expect(req.specification.fields).toHaveLength(1)
+    expect(req.specification.fields[0]).toEqual({
+      field: 'GeoY',
+      datasetField: 'point'
+    })
+  })
+
+  it('handles invalid coordinate values', () => {
+    const req = {
+      specification: {
+        fields: [
+          { field: 'GeoX', invalid: true },
+          { field: 'GeoY', invalid: true }
+        ]
+      },
+      fieldMappings: []
+    }
+    const res = {}
+    const next = vi.fn()
+
+    addDatabaseFieldToSpecification(req, res, next)
+    expect(req.specification.fields).toHaveLength(2)
+    expect(req.specification.fields[0]).toEqual({
+      field: 'GeoX',
+      datasetField: 'point',
+      invalid: true
+    })
+    expect(req.specification.fields[1]).toEqual({
+      field: 'GeoY',
+      datasetField: 'point',
+      invalid: true
+    })
+  })
+
+  it('handles fields with no matching field mapping', () => {
+    const req = {
+      specification: {
+        fields: [
+          { field: 'unknownField' }
+        ]
+      },
+      fieldMappings: []
+    }
+    const res = {}
+    const next = vi.fn()
+
+    addDatabaseFieldToSpecification(req, res, next)
+    expect(req.specification.fields).toHaveLength(1)
+    expect(req.specification.fields[0]).toEqual({
+      field: 'unknownField',
+      datasetField: 'unknownField'
+    })
+  })
+
+  it('calls next function', () => {
+    const req = {
+      specification: {
+        fields: [
+          { field: 'name' }
+        ]
+      },
+      fieldMappings: [
+        { field: 'name', replacement_field: 'full_name' }
+      ]
+    }
+    const res = {}
+    const next = vi.fn()
+
+    addDatabaseFieldToSpecification(req, res, next)
+    expect(next).toHaveBeenCalledTimes(1)
+  })
+})
+
+describe('getUniqueDatasetFieldsFromSpecification', () => {
+  it('gets unique dataset fields from specification', () => {
+    const req = {
+      specification: {
+        fields: [
+          { field: 'foo', datasetField: 'foo_field' },
+          { field: 'bar', datasetField: 'bar_field' },
+          { field: 'baz', datasetField: 'foo_field' } // duplicate
         ]
       }
-      const res = {}
-      const next = vi.fn()
+    }
+    const res = {}
+    const next = vi.fn()
 
-      addDatabaseFieldToSpecification(req, res, next)
-      expect(req.specification.fields).toHaveLength(2)
-      expect(req.specification.fields[0]).toEqual({
-        field: 'name',
-        datasetField: 'full_name'
-      })
-      expect(req.specification.fields[1]).toEqual({
-        field: 'address',
-        datasetField: 'physical_address'
-      })
-    })
+    getUniqueDatasetFieldsFromSpecification(req, res, next)
 
-    it('handles special case for GeoX and GeoY fields', () => {
-      const req = {
-        specification: {
-          fields: [
-            { field: 'GeoX' },
-            { field: 'GeoY' }
-          ]
-        },
-        fieldMappings: []
+    expect(req.uniqueDatasetFields).toEqual(['foo_field', 'bar_field'])
+    expect(next).toHaveBeenCalledTimes(1)
+  })
+
+  it('returns an empty array when specification.fields is empty', () => {
+    const req = {
+      specification: {
+        fields: []
       }
-      const res = {}
-      const next = vi.fn()
+    }
+    const res = {}
+    const next = vi.fn()
 
-      addDatabaseFieldToSpecification(req, res, next)
-      expect(req.specification.fields).toHaveLength(2)
-      expect(req.specification.fields[0]).toEqual({
-        field: 'GeoX',
-        datasetField: 'point'
-      })
-      expect(req.specification.fields[1]).toEqual({
-        field: 'GeoY',
-        datasetField: 'point'
-      })
-    })
+    getUniqueDatasetFieldsFromSpecification(req, res, next)
 
-    it('handles single GeoX field without GeoY', () => {
-      const req = {
-        specification: {
-          fields: [
-            { field: 'GeoX' }
-          ]
-        },
-        fieldMappings: []
-      }
-      const res = {}
-      const next = vi.fn()
+    expect(req.uniqueDatasetFields).toEqual([])
+    expect(next).toHaveBeenCalledTimes(1)
+  })
 
-      addDatabaseFieldToSpecification(req, res, next)
-      expect(req.specification.fields).toHaveLength(1)
-      expect(req.specification.fields[0]).toEqual({
-        field: 'GeoX',
-        datasetField: 'point'
-      })
-    })
+  it('throws an error when specification is not provided', () => {
+    const req = {}
+    const res = {}
+    const next = vi.fn()
 
-    it('handles single GeoY field without GeoX', () => {
-      const req = {
-        specification: {
-          fields: [
-            { field: 'GeoY' }
-          ]
-        },
-        fieldMappings: []
-      }
-      const res = {}
-      const next = vi.fn()
-
-      addDatabaseFieldToSpecification(req, res, next)
-      expect(req.specification.fields).toHaveLength(1)
-      expect(req.specification.fields[0]).toEqual({
-        field: 'GeoY',
-        datasetField: 'point'
-      })
-    })
-
-    it('handles invalid coordinate values', () => {
-      const req = {
-        specification: {
-          fields: [
-            { field: 'GeoX', invalid: true },
-            { field: 'GeoY', invalid: true }
-          ]
-        },
-        fieldMappings: []
-      }
-      const res = {}
-      const next = vi.fn()
-
-      addDatabaseFieldToSpecification(req, res, next)
-      expect(req.specification.fields).toHaveLength(2)
-      expect(req.specification.fields[0]).toEqual({
-        field: 'GeoX',
-        datasetField: 'point',
-        invalid: true
-      })
-      expect(req.specification.fields[1]).toEqual({
-        field: 'GeoY',
-        datasetField: 'point',
-        invalid: true
-      })
-    })
-
-    it('handles fields with no matching field mapping', () => {
-      const req = {
-        specification: {
-          fields: [
-            { field: 'unknownField' }
-          ]
-        },
-        fieldMappings: []
-      }
-      const res = {}
-      const next = vi.fn()
-
-      addDatabaseFieldToSpecification(req, res, next)
-      expect(req.specification.fields).toHaveLength(1)
-      expect(req.specification.fields[0]).toEqual({
-        field: 'unknownField',
-        datasetField: 'unknownField'
-      })
-    })
-
-    it('calls next function', () => {
-      const req = {
-        specification: {
-          fields: [
-            { field: 'name' }
-          ]
-        },
-        fieldMappings: [
-          { field: 'name', replacement_field: 'full_name' }
-        ]
-      }
-      const res = {}
-      const next = vi.fn()
-
-      addDatabaseFieldToSpecification(req, res, next)
-      expect(next).toHaveBeenCalledTimes(1)
-    })
+    expect(() => getUniqueDatasetFieldsFromSpecification(req, res, next)).toThrowError('specification is required')
   })
 })
